backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0

# Supabase client
supabase==2.3.4

# HTTP client for external API calls
httpx>=0.24.0,<0.25.0

# Data validation and settings
pydantic==2.5.0
pydantic-settings==2.1.0

# Environment variables
python-dotenv==1.0.0

# JSON serialization (faster than standard library)
orjson==3.9.10

# Retry logic for external services
tenacity==8.2.3

# Optional: OpenAI for AI features
openai==1.3.7
openpyxl==3.1.2

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx>=0.24.0,<0.25.0  # For testing async clients
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Additional dependencies for normalization script
asyncpg==0.29.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# CORS middleware
python-multipart==0.0.6
